% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_result.R
\name{is_valid}
\alias{is_valid}
\title{Is the validation result valid?}
\usage{
is_valid(x, ...)
}
\arguments{
\item{x}{A \code{js_validation_result}.}

\item{...}{Unused.}
}
\value{
\code{TRUE} when validation succeeded (no error-severity problems).
}
\description{
Is the validation result valid?
}
