% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_result.R
\name{js_error_messages}
\alias{js_error_messages}
\title{Error messages from a validation result}
\usage{
js_error_messages(x)
}
\arguments{
\item{x}{A \code{js_validation_result}.}
}
\value{
A character vector of messages whose severity is \code{error}.
}
\description{
Error messages from a validation result
}
