\name{jsonstructure_version}
\alias{jsonstructure_version}
\title{Package version}
\usage{
jsonstructure_version()
}
\value{
  The installed package version as a character string.
}
\description{
  Returns the installed \code{jsonstructure} package version.
}
\examples{
\dontrun{
jsonstructure_version()
}
}
